// becapbi/src/models/dataFamiliar.d.ts
type FormValues = {
  en_contacto_padres: boolean
  tiene_apoderado: boolean
  direccion_padres?: string
  telefono_padres?: string
  celular_padres?: string
  referencia_padres?: string
  tipo_vivienda_pad?: string
  nombres_apellido_apoderado?: string
  direccion_apoderado?: string
  telefono_apoderado?: string
  celular_apoderado?: string
  referencia_apoderado?: string
}

export default FormValues